+++ netbox_fms/tables.py
@@ -42,6 +42,8 @@ class FiberCableTypeTable(NetBoxTable):
     is_armored = columns.BooleanColumn(verbose_name=_("Armored"))
     deployment = tables.Column(verbose_name=_("Deployment"))
     fire_rating = tables.Column(verbose_name=_("Fire Rating"))
+    outer_diameter = tables.Column(verbose_name=_("Outer Diameter (mm)"))
+    twist_factor_ratio = tables.Column(verbose_name=_("Twist Factor"))
     instance_count = columns.LinkedCountColumn(
         viewname="plugins:netbox_fms:fibercable_list",
         url_params={"fiber_cable_type_id": "pk"},
@@ -60,6 +62,8 @@ class FiberCableTypeTable(NetBoxTable):
             "construction",
             "fiber_type",
             "strand_count",
+            "outer_diameter",
+            "twist_factor_ratio",
             "sheath_material",
             "jacket_color",
             "is_armored",